Extract system tests into separate files
