+++ test/system/importmap-laravel-eager.test.ts
@@ -1,0 +1,24 @@
+import { describe, test, expect } from "vitest"
+import { setupProject } from "../helpers/setup"
+
+const project = setupProject("importmap-laravel-eager")
+
+describe("System", () => {
+  test("importmap-laravel-eager", async () => {
+    expect(project.controllersFile).toBeUndefined()
+    expect(project.applicationFile).toBeUndefined()
+    expect(project.registeredControllers.length).toEqual(0)
+
+    await project.initialize()
+
+    expect(project.controllersFile).toBeDefined()
+    expect(project.relativePath(project.controllersFile.path)).toEqual("resources/js/controllers/index.js")
+
+    expect(project.applicationFile).toBeDefined()
+    expect(project.relativePath(project.applicationFile.path)).toEqual("resources/js/libs/stimulus.js")
+
+    expect(project.registeredControllers.length).toEqual(1)
+    expect(project.registeredControllers.map(controller => [controller.identifier, controller.loadMode])).toEqual([["hello", "stimulus-loading-eager"]])
+    expect(Array.from(project.controllerRoots)).toEqual(["resources/js/controllers"])
+  })
+})
